requirements.txt
tqdm==4.36.1
mypy==0.730
mypy-extensions==0.4.2
typing-extensions==3.7.4
overrides==2.0
numpy==1.17.2
tensorboard==2.0.0
tensorboardX==1.9
tensorflow-estimator==2.0.0
matplotlib==3.1.1
